Use context manager in unittests
Improved InvalidMessageDefinition error message for incorrect message size
